requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
celery>=5.3
kombu>=5.3
amqp>=5.2
requests>=2.32
python-dotenv>=1.0